move edit from bracket page to event page
